work/requirements.txt
Flask==1.1.0
numpy==1.16.2
opencv_python==4.1.0.25
requests==2.21.0
paddlepaddle==1.6.1
Twisted==19.7.0
Pillow==7.0.0
ConfigParser